requirements.txt
fastapi
uvicorn[standard]
pandas
pinecone-client==2.2.4
python-dotenv
requests
tiktoken
pydantic